Some more restructuring
