+++ judge.py
@@ -8,7 +8,7 @@ import threading
 
 from custom_logger import main_logger
 from protocol import Connection
-from protocol.judge import Commands, JudgeProtocol
+from protocol.judge import Commands, Protocol
 
 HOST = "localhost"
 PORT = 12345
@@ -34,7 +34,7 @@ def _handle_connections(client_socket: socket.socket, addr: tuple[str, int]):
         logger.info(
             f"Checking if the runner with IP {ip} on port {port} is initialized correctly..."
         )
-        JudgeProtocol.send_command(connection, Commands.CHECK)
+        Protocol.send_command(connection, Commands.CHECK)
         logger.info(f"Runner with IP {ip} on port {port} initialized.")
 
     except socket.timeout:
